lambdas/ml_export/requirements.txt
snowflake-connector-python==3.12.3
